backend/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
matplotlib